Langchain_Agents/requirements.txt
langchain
python-dotenv
ipykernel
langchain_community
pypdf
bs4
langchain_huggingface
langchain_chroma
arxiv
pymupdf
wikipedia
langchain_text_splitters
langchain_openai
chromadb
sentence_transformers
faiss_cpu
langchain_ollama
groq
mysql-connector-python
sqlalchemy
fastapi
uvicorn
langserve
sse-starlette
psycopg2
snowflake
streamlit
pymysql
langchain-openai
